rename blockgen->walgen
- to make room for new blockgen
