+++ cmd/thanosbench/walgen.go
@@ -6,18 +6,18 @@ import (
 	"github.com/go-kit/kit/log"
 	"github.com/oklog/run"
 	"github.com/thanos-io/thanos/pkg/extflag"
-	"github.com/thanos-io/thanosbench/pkg/blockgen"
+	"github.com/thanos-io/thanosbench/pkg/walgen"
 	"gopkg.in/alecthomas/kingpin.v2"
 )
 
-func registerBlockgen(m map[string]setupFunc, app *kingpin.Application) {
-	cmd := app.Command("blockgen", "")
+func registerWalgen(m map[string]setupFunc, app *kingpin.Application) {
+	cmd := app.Command("walgen", "Generates TSDB data into WAL files.")
 	config := extflag.RegisterPathOrContent(cmd, "config", "JSON file for series config", true)
 
 	outputDir := cmd.Flag("output-dir", "Output directory for generated TSDB data.").Required().String()
 
 	// TODO(bwplotka): Consider mode in which it generates the data only if empty work dir.
-	m["blockgen"] = func(g *run.Group, logger log.Logger) error {
+	m["walgen"] = func(g *run.Group, logger log.Logger) error {
 		g.Add(func() error {
 			configContent, err := config.Content()
 			if err != nil {
@@ -27,7 +27,7 @@ func registerBlockgen(m map[string]setupFunc, app *kingpin.Application) {
 				return err
 			}
 
-			return blockgen.GenerateTSDB(logger, *outputDir, configContent)
+			return walgen.GenerateTSDB(logger, *outputDir, configContent)
 		}, func(error) {})
 		return nil
 	}
