+++ pkg/walgen/walgen_test.go
@@ -1,4 +1,4 @@
-package blockgen
+package walgen
 
 import (
 	"testing"
